tests: Create mocks for services, fix service tests
