+++ src/app/services/auth/auth.service.fake.ts
@@ -16,50 +16,26 @@ export interface IAuthService {
 }
 
 @Injectable()
-export class AuthService implements IAuthService {
+export class MockAuthService implements IAuthService {
   key = 'tcp-angular'
   jwtHelper = new JwtHelperService()
 
-  constructor(private http: HttpClient, private router: Router) {}
+  constructor(private router: Router) {}
 
-  login(username: string, password: string) {
-    this.http
-      .get('login', {
-        headers: new HttpHeaders({ u: username, p: password }),
-      })
-      .subscribe(
-        (data: string) => {
-          localStorage.setItem(this.key, data)
-          this.router.navigateByUrl('')
-        },
-        (err: HttpErrorResponse) => {
-          alert(err.error)
-        }
-      )
-  }
+  login(username: string, password: string) {  }
 
   logout() {
-    localStorage.removeItem(this.key)
     this.router.navigateByUrl('login')
   }
 
   getToken(decoded: boolean = false) {
-    const token = localStorage.getItem(this.key)
-    try {
-      if (token) {
-        const decodedToken = this.jwtHelper.decodeToken(token)
-        if (decodedToken.exp - new Date().getTime() < 0) {
-          this.logout()
-          return null
-        }
+    if (decoded) {
+      return {
+        email: 'jon.doe@gmail.com',
+        role: Role.admin,
       }
-      if (decoded) {
-        return this.jwtHelper.decodeToken(token)
-      } else {
-        return token
-      }
-    } catch (err) {
-      this.logout()
+    } else {
+      return {}
     }
   }
