+++ src/app/services/auth/auth.service.spec.ts
@@ -1,9 +1,14 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing'
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { AuthService } from './auth.service'
 
 describe('AuthService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}))
+  beforeEach(() => TestBed.configureTestingModule({
+    providers: [AuthService],
+    imports: [HttpClientTestingModule, RouterTestingModule]
+  }))
 
   it('should be created', () => {
     const service: AuthService = TestBed.get(AuthService)
