+++ src/app/services/employees/employees.service.spec.ts
@@ -1,9 +1,13 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing'
 
 import { EmployeesService } from './employees.service'
 
 describe('EmployeesService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}))
+  beforeEach(() => TestBed.configureTestingModule({
+    providers: [EmployeesService],
+    imports: [HttpClientTestingModule]
+  }))
 
   it('should be created', () => {
     const service: EmployeesService = TestBed.get(EmployeesService)
