+++ src/app/services/skills/skills.service.ts
@@ -1,74 +1,38 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core'
 import { BehaviorSubject } from 'rxjs'
 
 import { ISkill } from '../../models/skill.interface'
-import { dummySkillCategories } from '../skill-categories/skill-categories.service'
+import { dummySkills } from './skills.service.fake'
 
-export const dummySkills: ISkill[] = [
-  {
-    id: 1,
-    name: 'Java',
-    category: dummySkillCategories.find(c => c.name === 'Software'),
-  },
-  {
-    id: 2,
-    name: 'JavaScript',
-    category: dummySkillCategories.find(c => c.name === 'Software'),
-  },
-  {
-    id: 3,
-    name: 'Business Analysis',
-    category: dummySkillCategories.find(c => c.name === 'Business'),
-  },
-  {
-    id: 4,
-    name: 'Business Development',
-    category: dummySkillCategories.find(c => c.name === 'Business'),
-  },
-  {
-    id: 5,
-    name: 'Recruiting',
-    category: dummySkillCategories.find(c => c.name === 'Business'),
-  },
-  {
-    id: 6,
-    name: 'Scrum Master',
-    category: dummySkillCategories.find(c => c.name === 'Agile'),
-  },
-  {
-    id: 7,
-    name: 'Agile Coach',
-    category: dummySkillCategories.find(c => c.name === 'Agile'),
-  },
-  {
-    id: 8,
-    name: 'Jenkins',
-    category: dummySkillCategories.find(c => c.name === 'Dev Ops'),
-  },
-  { id: 9, name: 'AWS', category: dummySkillCategories.find(c => c.name === 'Dev Ops') },
-]
+export interface ISkillsService {
+  readonly list: BehaviorSubject<ISkill[]>
+  fetch(): void
+  addSkill(skill: ISkill): void
+  updateSkill(skill: ISkill): void
+  deleteSkill(id: number): void
+}
 
-@Injectable({
-  providedIn: 'root',
-})
-export class SkillsService {
+@Injectable()
+export class SkillsService implements ISkillsService{
   readonly list = new BehaviorSubject<ISkill[]>([])
 
-  constructor() {}
+  constructor(private http: HttpClient) {}
 
   fetch(): void {
     this.list.next(dummySkills)
   }
 
   addSkill(skill: ISkill) {
-    let newList: ISkill[]
-    if (skill.id) {
-      this.deleteSkill(skill.id)
-      newList = this.list.value
-    } else {
-      newList = this.list.value
-      skill.id = newList.length
-    }
+    const newList = this.list.value
+    skill.id = newList.length
+    newList.push(skill)
+    this.list.next(newList)
+  }
+
+  updateSkill(skill: ISkill) {
+    this.deleteSkill(skill.id)
+    const newList = this.list.value
     newList.push(skill)
     this.list.next(newList)
   }
